unittest: Ensure different timestamps on generated files.
The test P2PManagerTest.Housekeeping needs to generate files with
different timestamps to test the Housekeeping function. These
timestamps are affected by the file system's time resolution which
is 1 second on ext2/ext3 and 1 nanosecond on ext4. The previous fix
introduced by CL:167882 didn't solve the problem because touch -t
replaces the "Modify" and "Create" times, but not the "Change" time
which is the one used by Housekeeping.

This patch ensures that the timestamps on the created files are
different sleeping enough time to allow the kernel put a different
timestamp for the "Change" time. The slept time depends on the
file system time resolution, making it shorter when ran on
file systems with higher resolution.

BUG=chromium:290257,chromium:281694
TEST=Run unittest on an environment where /tmp is an ext2 FS.

Change-Id: I5631ef317813b83443b36e0e5b660cafb8616f59
Reviewed-on: https://chromium-review.googlesource.com/169381
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>