PolicyManager: Initial classes for PM variables.
The PolicyManager code will reside on the new sub-directory
policy_manager/ to reduce the accidental coupling of this code with
other parts of the Update Engine. The subdirectory '.' is included
in the include path to allow code in subdirectories use the full
path such as "policy_manager/variable.h" when including files
instead of a relative path.

This first patch adds the basic interface for PM variables to be
exposed by PM providers. The patch includes the Variable class.

BUG=None
TEST=None, only interfaces were added.

Change-Id: I16828ba08a1ccf7f318b986b718319bbacea59ee
Reviewed-on: https://chromium-review.googlesource.com/181536
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>