PM: Add stock policy_manager.conf file.
The idea is that this file can be provided by a board-specific
package. To make it easy for such developers to figure out what they
can tweak, provide a helpful, well-documented configuration file.

Another CL will modify the update_engine ebuild so that it gets
installed (in /etc) along with the other files needed.

BUG=chromium:352739
TEST=None

Change-Id: I3d1045dbb9f5fd5d3ba31ef43c50addcf824fc96
Reviewed-on: https://chromium-review.googlesource.com/196700
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>