PM: Fix PolicyManager calls with multiple variadic template arguments.
We've been using a pattern among PolicyManager calls whereas a single
variadic template argument was used to describe the types of first-class
arguments to the template method, but also the type signature of
a method that is another argument. This was causing issues where
template type inference was unable to resolve automatic casting / type
promotion among these two (otherwise compatible) type lists.

Instead, we now use a separate variadic template argument for these two
instances: ExpectedArgs... represents the types of arguments expected by
the policy methods, whereas ActualArgs... represents the types of
arguments that are actually provided to the PolicyManager. This allows
template type inference to succeed. To ensure that type
incompatibilities do not propagate far, resulting in cryptic compiler
errors, we further use explicit instantiation of subsequent templates in
the two public methods, PolicyREquest and AsyncPolicyRequest. This
ensures early on that the actual arguments can be converted to the
expected types.

Also, this removes parameter names from the declaration of function
arguments; they are not needed and decrease clarity.

BUG=None
TEST=Unit tests.

Change-Id: Ia5552714477c3f30f6ff1717461ba4461bea38be
Reviewed-on: https://chromium-review.googlesource.com/200570
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>