Fix ServerDiesTest and FailureTest which depended on you running apache.
This CL fixes the issue that URL in both of these tests defaulted to
no-port which meaned the tests were going to http://localhost.
On some machines, that might point to an existing server running
on port 80 (for me it was apache2 configured by
autotest). This CL changes FailureTest to use a different
hostname that should never match. In the case of the ServerDiesTest,
we use the correct port of the server we kill during the test.

I've also added documentation to what these tests should actually do.

BUG=chromium:383200
TEST=Ran unittests with and without apache up.

Change-Id: I6b4168eada48dcd59fad23ed7ff6577b18e4ef81
Reviewed-on: https://chromium-review.googlesource.com/203347
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>